+++ .github/workflows/ci-actions.yml
@@ -21,3 +21,8 @@ jobs:
 
       - name: unit test
         run: npm run test
+        
+      - name: Upload coverage reports to Codecov
+        uses: codecov/codecov-action@v3
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
