+++ package.json
@@ -8,7 +8,7 @@
     "start": "next start",
     "lint": "next lint",
     "test": "jest --transformIgnorePatterns '/node_modules/(?!firebase)/'",
-    "test:coverage": "jest --coverage",
+    "test:coverage": "jest --coverage && codecov",
     "build:analyze": "ANALYZE=true yarn build",
     "vercel-build": "next build",
     "docker:build": "docker build -t my-news-vault .",
@@ -31,6 +31,7 @@
   },
   "devDependencies": {
     "@next/bundle-analyzer": "^13.4.5",
+    "@playwright/test": "^1.38.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "@types/jest": "^29.5.1",
@@ -41,6 +42,7 @@
     "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.59.2",
     "@typescript-eslint/parser": "^5.59.2",
+    "codecov": "^3.8.3",
     "encoding": "^0.1.13",
     "eslint": "^8.39.0",
     "eslint-config-airbnb": "^19.0.4",
