+++ src/msw/handlers/newsScrapHandlers.ts
@@ -1,0 +1,1 @@
+export default [];
