+++ src/msw/server.ts
@@ -1,5 +1,5 @@
 import { setupServer } from 'msw/node';
-import { handlers } from './handlers';
+import { handlers } from './rootHandler';
 
 // mock server 생성
 export const server = setupServer(...handlers);
